Ship beta launch package with Flutter app, FCM, and ops tooling.
Un-ignores database migrations/seeders, adds create-business and language UX, and hardens login routing for local beta installs.

+++ app/Support/Analytics/AnalyticsDateRange.php
@@ -107,7 +107,7 @@ final class AnalyticsDateRange
 
     public function previousPeriod(): self
     {
-        $days = $this->fromLocal->startOfDay()->diffInDays($this->toLocal->startOfDay()) + 1;
+        $days = (int) ($this->fromLocal->startOfDay()->diffInDays($this->toLocal->startOfDay()) + 1);
         $previousEnd = $this->fromLocal->subDay()->endOfDay();
         $previousStart = $previousEnd->subDays($days - 1)->startOfDay();
